bugfixes, todo integrate ssl_context.generate_ssl_context into httpy.py|
